Move: @component/list 내부에 디렉토리를 만들어서 같은 종류의 컴포넌트끼리 정리함. shipping 관련 컴포넌트는 @component/shipping 컴포넌트로 이동.
