+++ client/src/page/admin/photo/list/Index.tsx
@@ -6,7 +6,7 @@ import { faPlus, faArrowRight } from '@fortawesome/free-solid-svg-icons';
 import Card from '@component/card/basic/Card';
 import CardBody from '@component/card/basic/CardBody';
 import Button from '@component/form/Button';
-import PhotoListWithFilter from '@component/list/PhotoListWithFilter';
+import PhotoListWithFilter from '@component/list/photo/PhotoListWithFilter';
 import './Index.scss';
 
 function PhotoListPage() {
