+++ client/src/page/admin/voucher/writer/content/VoucherSection.tsx
@@ -8,7 +8,7 @@ import Button from '@component/form/Button';
 import Card from '@component/card/basic/Card';
 import CardHeader from '@component/card/basic/CardHeader';
 import CardBody from '@component/card/basic/CardBody';
-import PhotoListWithFilter from '@component/list/PhotoListWithFilter';
+import PhotoListWithFilter from '@component/list/photo/PhotoListWithFilter';
 import PhotoList from './photo/PhotoList';
 import { State, Action } from '../reducer';
